chore: convert example to typescript
