+++ IapExample/index.tsx
@@ -1,9 +1,5 @@
-/**
- * @format
- */
-
 import {AppRegistry} from 'react-native';
-import App from './App';
 import {name as appName} from './app.json';
+import {App} from './src/App';
 
 AppRegistry.registerComponent(appName, () => App);
